:bug: fix: #26 useSearchParams 사용으로 인한 CSR 오류 수정 (Suspense로 감싸도록 변경)
