+++ src/app/story/page.tsx
@@ -1,23 +1,12 @@
-'use client';
+import React, { Suspense } from 'react';
+import StoryListPage from '@components/story/StoryListPage';
 
-import React from 'react';
-import { useSearchParams } from 'next/navigation';
-import StoryHeroSection from '@components/story/sections/StoryHeroSection';
-import StoryListSection from '@components/story/sections/StoryListSection';
-import StorySearchSection from '@components/story/sections/StorySearchSection';
-
-function StoryListPage() {
-  const searchParams = useSearchParams();
-  const keyword = searchParams.get('q') ?? '';
-  const region = searchParams.get('region') ?? '';
-
+function Page() {
   return (
-    <section>
-      <StoryHeroSection />
-      <StorySearchSection keyword={keyword} />
-      <StoryListSection keyword={keyword} region={region} />
-    </section>
+    <Suspense fallback={<div>로딩 중...</div>}>
+      <StoryListPage />
+    </Suspense>
   );
 }
 
-export default StoryListPage;
+export default Page;
